+++ alpharaw/bruker/spectrum.py
@@ -1,0 +1,260 @@
+import numpy as np
+
+from alpharaw.utils.pjit import njit, pjit
+
+
+
+@pjit(
+    # signature_or_function="void(i8,i8[:],i8[:],i8[:],u4[:],u2[:],u4[:],f8[:],i8[:],i8[:])"
+)
+def set_precursor(
+    precursor_index: int,
+    offset_order: np.ndarray,
+    precursor_offsets: np.ndarray,
+    quad_indptr: np.ndarray,
+    tof_indices: np.ndarray,
+    intensities: np.ndarray,
+    spectrum_tof_indices: np.ndarray,
+    spectrum_intensity_values: np.ndarray,
+    spectrum_indptr: np.ndarray,
+    spectrum_counts: np.ndarray,
+) -> None:
+    """Sum the intensities of all pushes belonging to a single precursor.
+
+    IMPORTANT NOTE: This function is decorated with alphatims.utils.pjit.
+    The first argument is thus expected to be provided as an iterable
+    containing ints instead of a single int.
+
+    Parameters
+    ----------
+    precursor_index : int
+        The precursor index indicating which MS2 spectrum to determine.
+    offset_order : np.int64[:]
+        The order of self.precursor_indices, obtained with np.argsort.
+    precursor_offsets : np.int64[:]
+        An index pointer array for precursor offsets.
+    quad_indptr : np.int64[:]
+        The self.quad_indptr array of a TimsTOF object.
+    tof_indices : np.uint32[:]
+        The self.tof_indices array of a TimsTOF object.
+    intensities : np.uint16[:]
+        The self.intensity_values array of a TimsTOF object.
+    spectrum_tof_indices : np.uint32[:]
+        A buffer array to store tof indices of the new spectrum.
+    spectrum_intensity_values : np.float64[:]
+        A buffer array to store intensity values of the new spectrum.
+    spectrum_indptr : np.int64[:]
+        An index pointer array defining the original spectrum boundaries.
+    spectrum_counts : np. int64[:]
+        An buffer array defining how many distinct tof indices the new
+        spectrum has.
+    """
+    offset = spectrum_indptr[precursor_index]
+    precursor_offset_lower = precursor_offsets[precursor_index]
+    precursor_offset_upper = precursor_offsets[precursor_index + 1]
+    selected_offsets = offset_order[
+        precursor_offset_lower: precursor_offset_upper
+    ]
+    starts = quad_indptr[selected_offsets]
+    ends = quad_indptr[selected_offsets + 1]
+    offset_index = offset
+    for start, end in zip(starts, ends):
+        spectrum_tof_indices[
+            offset_index: offset_index + end - start
+        ] = tof_indices[start: end]
+        spectrum_intensity_values[
+            offset_index: offset_index + end - start
+            ] = intensities[start: end]
+        offset_index += end - start
+    offset_end = spectrum_indptr[precursor_index + 1]
+    order = np.argsort(spectrum_tof_indices[offset: offset_end])
+    current_index = offset - 1
+    previous_tof_index = -1
+    for tof_index, intensity in zip(
+        spectrum_tof_indices[offset: offset_end][order],
+        spectrum_intensity_values[offset: offset_end][order],
+    ):
+        if tof_index != previous_tof_index:
+            current_index += 1
+            spectrum_tof_indices[current_index] = tof_index
+            spectrum_intensity_values[current_index] = intensity
+            previous_tof_index = tof_index
+        else:
+            spectrum_intensity_values[current_index] += intensity
+    spectrum_tof_indices[current_index + 1: offset_end] = 0
+    spectrum_intensity_values[current_index + 1: offset_end] = 0
+    spectrum_counts[precursor_index] = current_index + 1 - offset
+
+
+@pjit
+def centroid_spectra(
+    index: int,
+    spectrum_indptr: np.ndarray,
+    spectrum_counts: np.ndarray,
+    spectrum_tof_indices: np.ndarray,
+    spectrum_intensity_values: np.ndarray,
+    window_size: int,
+):
+    """Smoothen and centroid a profile spectrum (inplace operation).
+
+    IMPORTANT NOTE: This function will overwrite all input arrays.
+
+    IMPORTANT NOTE: This function is decorated with alphatims.utils.pjit.
+    The first argument is thus expected to be provided as an iterable
+    containing ints instead of a single int.
+
+    Parameters
+    ----------
+    index : int
+        The push index whose intensity_values and tof_indices will be
+        centroided.
+    spectrum_indptr : np.int64[:]
+        An index pointer array defining the (untrimmed) spectrum boundaries.
+    spectrum_counts : np. int64[:]
+        The original array defining how many distinct tof indices each
+        spectrum has.
+    spectrum_tof_indices : np.uint32[:]
+        The original array containing tof indices.
+    spectrum_intensity_values : np.float64[:]
+        The original array containing intensity values.
+    window_size : int
+        The window size to use for smoothing and centroiding peaks.
+    """
+    start = spectrum_indptr[index]
+    end = start + spectrum_counts[index]
+    if start == end:
+        return
+    mzs = spectrum_tof_indices[start: end]
+    ints = spectrum_intensity_values[start: end]
+    smooth_ints = ints.copy()
+    for i, self_mz in enumerate(mzs[:-1]):
+        for j in range(i + 1, len(mzs)):
+            other_mz = mzs[j]
+            diff = other_mz - self_mz + 1
+            if diff >= window_size:
+                break
+            smooth_ints[i] += ints[j] / diff
+            smooth_ints[j] += ints[i] / diff
+    pre_apex = True
+    maxima = [mzs[0]]
+    intensities = [ints[0]]
+    for i, self_mz in enumerate(mzs[1:], 1):
+        if self_mz > mzs[i - 1] + window_size:
+            maxima.append(mzs[i])
+            intensities.append(0)
+            pre_apex = True
+        elif pre_apex:
+            if smooth_ints[i] < smooth_ints[i - 1]:
+                pre_apex = False
+                maxima[-1] = mzs[i - 1]
+        elif smooth_ints[i] > smooth_ints[i - 1]:
+            maxima.append(mzs[i])
+            intensities.append(0)
+            pre_apex = True
+        intensities[-1] += ints[i]
+    spectrum_tof_indices[start: start + len(maxima)] = np.array(
+        maxima,
+        dtype=spectrum_tof_indices.dtype
+    )
+    spectrum_intensity_values[start: start + len(maxima)] = np.array(
+        intensities,
+        dtype=spectrum_intensity_values.dtype
+    )
+    spectrum_counts[index] = len(maxima)
+
+
+
+@pjit
+def filter_spectra_by_abundant_peaks(
+    index: int,
+    spectrum_indptr: np.ndarray,
+    spectrum_counts: np.ndarray,
+    spectrum_tof_indices: np.ndarray,
+    spectrum_intensity_values: np.ndarray,
+    keep_n_most_abundant_peaks: int,
+):
+    """Filter a spectrum to retain only the most abundant peaks.
+
+    IMPORTANT NOTE: This function will overwrite all input arrays.
+
+    IMPORTANT NOTE: This function is decorated with alphatims.utils.pjit.
+    The first argument is thus expected to be provided as an iterable
+    containing ints instead of a single int.
+
+    Parameters
+    ----------
+    index : int
+        The push index whose intensity_values and tof_indices will be
+        centroided.
+    spectrum_indptr : np.int64[:]
+        An index pointer array defining the (untrimmed) spectrum boundaries.
+    spectrum_counts : np. int64[:]
+        The original array defining how many distinct tof indices each
+        spectrum has.
+    spectrum_tof_indices : np.uint32[:]
+        The original array containing tof indices.
+    spectrum_intensity_values : np.float64[:]
+        The original array containing intensity values.
+    keep_n_most_abundant_peaks : int
+        Keep only this many abundant peaks.
+    """
+    start = spectrum_indptr[index]
+    end = start + spectrum_counts[index]
+    if end - start <= keep_n_most_abundant_peaks:
+        return
+    mzs = spectrum_tof_indices[start: end]
+    ints = spectrum_intensity_values[start: end]
+    selected_indices = np.sort(
+        np.argsort(ints)[-keep_n_most_abundant_peaks:]
+    )
+    count = len(selected_indices)
+    spectrum_tof_indices[start: start + count] = mzs[selected_indices]
+    spectrum_intensity_values[start: start + count] = ints[selected_indices]
+    spectrum_counts[index] = count
+
+
+@pjit
+def trim_spectra(
+    index: int,
+    spectrum_tof_indices: np.ndarray,
+    spectrum_intensity_values: np.ndarray,
+    spectrum_indptr: np.ndarray,
+    trimmed_spectrum_tof_indices: np.ndarray,
+    trimmed_spectrum_intensity_values: np.ndarray,
+    new_spectrum_indptr: np.ndarray,
+) -> None:
+    """Trim remaining bytes after merging of multiple pushes.
+
+    IMPORTANT NOTE: This function is decorated with alphatims.utils.pjit.
+    The first argument is thus expected to be provided as an iterable
+    containing ints instead of a single int.
+
+    Parameters
+    ----------
+    index : int
+        The push index whose intensity_values and tof_indices will be trimmed.
+    spectrum_tof_indices : np.uint32[:]
+        The original array containing tof indices.
+    spectrum_intensity_values : np.float64[:]
+        The original array containing intensity values.
+    spectrum_indptr : np.int64[:]
+        An index pointer array defining the original spectrum boundaries.
+    trimmed_spectrum_tof_indices : np.uint32[:]
+        A buffer array to store new tof indices.
+    trimmed_spectrum_intensity_values : np.float64[:]
+        A buffer array to store new intensity values.
+    new_spectrum_indptr : np.int64[:]
+        An index pointer array defining the trimmed spectrum boundaries.
+    """
+    start = spectrum_indptr[index]
+    new_start = new_spectrum_indptr[index]
+    new_end = new_spectrum_indptr[index + 1]
+    trimmed_spectrum_tof_indices[new_start: new_end] = spectrum_tof_indices[
+        start: start + new_end - new_start
+    ]
+    trimmed_spectrum_intensity_values[
+        new_start: new_end
+    ] = spectrum_intensity_values[
+        start: start + new_end - new_start
+    ]
+
